Build with clang-cl
The call-detection trait formed a pointer to member of the derived class
for an api inherited from Ioc, which is not an exact match for taking the
address of an overloaded function. clang rejected it, so unfold() took its
fallback overload and handed the engine API wrapper objects in place of
the native structs -- silently, since both overloads return T const&.

OpenMP now comes from the imported target: OpenMP_CXX_FLAGS is a single
space-separated string that clang-cl receives as one argument and ignores,
so OpenMP was disabled there.

The remaining fixes are constructs MSVC accepts: SFINAE on a condition
independent of the function template's parameters, a partial
specialization declaring uint64_t for a uint32_t primary template, an
attribute inside a decl-specifier-seq, narrowing in a GUID initializer,
and a '##' that pastes no valid token. Non-MSVC toolchains are now
rejected at configure time, as is a 32-bit one.

+++ engine/osinteraction/listener.h
@@ -46,7 +46,7 @@ protected:
 };
 
 
-template<uint64_t head_message, uint64_t ... tail_messages>
+template<uint32_t head_message, uint32_t ... tail_messages>
 class ConcreteListener<head_message, tail_messages...> : public ConcreteListener<tail_messages...>
 {
 protected:
